repair some synchronization testing
TestResynchronize was not a valuable test, it was removed. The test
involved manually removing a block, connecting without that block, and
then re-adding it. I could not figure out what that was covering, and
there were no comments to help. It's been under `t.Skip()` for a long
time, so I removed it altogether.