bring back more commented out testing
A bug was discovered where the hardfork was being applied improperly.
Unable to figure out the exact reason behind the problem, I took the
more obviously correct approach of copying the old behavior into a new
function, and then before the hardfork height calling the old function,
providing a guaranteed line-for-line correctness.

The new function is now not cluttered with any hardfork branching.