repair an nd-failing test
fixes #1258

There was a bug in the testing code where it tried to create bad storage
proofs. But, it created bad storage proofs by leaving the 'base' blank,
which would not succeed if the base was only 1 byte and the randomly
generate base had an emtpy value (1-in-256 chance for 1 byte).

This PR fixes that.