remove t.Parallel from networking tests (#1270)
The networking tests wait until a network action is complete, but have a
timeout in case the action doesn't complete in time. Running the tests
in parallel makes it much more likely that the tests don't complete in
time because other tests are hogging all of the cpu and disk resources.

For tests which have sensitive timeouts, t.Parallel() has been removed.