Retain storage obligations after submitting proof
Previously, the host would delete storage obligations after they had
been completed. Now, the host will retain them, albeit the sector roots
are deleted (to preserve space in the database).

There's also a new field, specifically for the termination of the
storage obligation. The final status (failed, succeeded, or rejected) is
kept in the storage obligation, so that it may easily be checked.