Restore shared pointer between allHosts and activeHosts
In the hostdb, there's an implicit assumption that the hostEntry in the
allHosts map is the same hostEntry in the activeHosts map. The
persistence was not being loaded to reflect this correctly.

This PR restores that. Better safety in the hostdb would solve all of
these edge cases, this will be addressed at some point.

fix #1281