store a status string in the storage obligations
When a storage obligation is completed, a status string is added to the
obligation stored on disk. This allows hosts to view their contracts
historically and see why they failed in the event of a failure, which
will help with debugging/troubleshooting

I'm not entirely certain that a string is the best move. An alternate
option has us using error codes. A string is nice because it lets you
see the actual meaning of the error directly in the database. But if we
ever change the errors, the errors could be lost from the codebase.

I think a string is more likely to make sense in the long run, but it
does mean you could be left with a string that you are unable to grep.