add thread safety to the host
The host's storage obligation management, especially the parts where the
host is removing storage obligations, are now being handled more safely,
which should prevent some of the major errors that have been witnessed
in the wild by renters.

To help with the testing, the host is now using a more full-featured
version of the thread-group, including using pausing to help with
testing.