better storage obligation safety
The host now locks storage obligations before ever pulling them from the
database, and keeps the locks until the host is no longer going to be
adding them to the database. This resolves some issues relating to
storage obligation corruption, which can cause errors like 'Consensus
Conflict' and can cause the host to report the wrong revision number, or
to have the wrong sector roots, etc.