ErrPotentialDoubleSpend -> ErrIncompleteTransaction
ErrPotentialDoubleSpend was never a good name. All it really means is
that there are multiple transactions builders open at the same time, and
while the wallet is not reporting a reduced unconfirmed balance (the
transaction is not complete), the wallet also doesn't have access to the
outputs that make up the balance.