remove safelocks from host
Very slow disk I/O was causing false positives in the deadlock detector
in the host. While it's a pretty significant issue that the I/O is that
slow, it's more significant that false deadlocks are being detected.

The deadlock detector should not be triggering due to I/O waits. But
until the situation with the locking + I/O can be resolved, we're going
to have to remove the safelocks.