wait for port forwarding on host shutdown
Host was previously not waiting for the port forwarding goroutine to
complete before proceeding to the AfterStop functions of the host, which
meant an error during managedClearPort could cause a panic in the
logger.

The fix is as simple as wrapping the goroutine in a thread group.

A regression test has been added to demonstrate that the problem has
been fixed. Adding the test resulted in a decent amount dependency code
needing to be injected into the codebase, but overall it's pretty clean.