Resolve some nondeterministic failures through flushing
The host was experiencing nondeterministic failures because storage
obligations were changed to be updated in goroutines, as to not block
the consensus. This means that, especially during highly parallel tests,
sometimes actions were no longer being completed before the test was
checking that the action had completed.

Application of flushing makes sure that the test is synchronizing
properly.