better concurrency patterns around gateway permanent threads
A new convention is introduced. Some threads are meant to be around
until the threadgroup is closed, which means that they cannot hold a
tg.Add call throughout their lifetime. These threads are different from
'threaded' functions, which will close themselves without a signal from
the threadgroup. The permanaent threads are distinguished by changing
the prefix from 'threaded' to 'permanent'. For the most part, they
should be treated the same as 'threaded' functions, except that they
need to use explicit signalling for shutdown, and cannot hold a
'tg.Add'.

threadedNodeManager was split into to functions. The first,
'permanentNodeManager' is responsible for adding peers from the node
list if the number of peers gets low. The second, 'permanaentNodePurger'
is responsible for purging nodes from the nodelist if they are offline.