shuffle around some gateway constants
I started migrating a lot of the gateway constants into a consts.go
file, and then adjusting some of them depending on the build type. A big
thing I had in mind was the Sia ant-farm, we want to be able to
reasonably hit some of the limits on the Sia ant-farm to verify that the
gateway is operating correctly when there are a lot of peers.

This will be much easier to do now that the constants are lower.