add gateway test for overloaded bootstrap
A new test was added that serves two functions. First it checks that
nodes trying to connect to a bootstrap will have success getting onto
the network even if the bootstrap node has more peers than it can
handle.

Second, it checks that in the event of a large node outage on the
network, the nodes will re-connect to eachother.

Partition tolerance in testing is fairly low, as the
'wellConnectedThreshold' is low.