add outbound peers even if they are already inbound peers
This does not violate any security assumptions, as the important thing
is that we chose the peer instead of the peer choosing us. It will help
certain operations complete faster.

Additionally, do not let an error in 'addNode' stop a peer connection.
An error in 'addNode' should prevent a peer from making it into the
nodelist, but if the peer connection was successful already then having
it not be a node is okay.