set conn deadline in Dial call, not on conn
The go std library will opportunistically reuse connections when it can.
If you set the deadline on the conn itself, it can cause a lot of i/o
timeout errors. If you set the deadline on the Dial call however, the
std library will extend the deadline as makes sense when reusing conns.