make editors/downloaders exclusive
previously, multiple editors and downloaders could be created for
the same contract, leading to desynchronization. This is mitigated
by adding a map to the contractor that tracks which contracts are
being revised. This is not intended to be a robust solution; it
would be better to interleave editing and downloading such that
neither is blocked.