add upload/download coordination
a simple pair of flags is used to indicate whether the renter
is uploading or downloading. This could probably be condensed
into a single flag if I were more clever, but I felt clarity
was more important. Also this code isn't intended to stick
around for very long (famous last words).