additional reorganization and bugfixes
Went through and read more of the code, gained a much stronger understanding of
what is happening and what should be happening, even enough to fix a few
mistakes and make a better logical construction. Some of the testing had to be
updated because I reordered the events of the sync loop, and the testing
depends on interrupting the sync loop to simulate failures (doing some parts
manually but not others).

At this point I've finished looking at the storage folder code that's in place,
the sector code and testing code still remains.