adjust constants given new capabilities
The renter is a lot more stable now, and has the ability to replace offline
hosts. This means we can safely reduce the redundacny from 6x to 4x, which will
help upload times. We have also increased the total number of hosts in use from
24 to 32, which will increase overall reliability relative to the redundancy.
Filanny, we have increased the default number of hosts to 42, which is 10 more
hosts than are needed to get a chunk to full redundancy. This means that the
renter can largely tolerate a lot of failing hosts without the upload being
impacted. It also means a greater percentage of hosts on the network will be
getting data, which will help with host encouragement.