Prune nodes more rapidly when there are many
The gateway was having a problem with a large node list, where many of the nodes
on that list were unconnectable, but the pruning was not happening fast enough
to keep the node list reasonably useful. Now, the gateway will prune nodes at a
more rapid rate when there are a lot of nodes.

This is an incomplete fix, we also need to extend the dial bit to do the full
handshake, merely seeing that a node is online is not enough if that node is not
running a recent version.