fix TestRenterRenew
Mining testPeriod blocks causes the test to fail because the renter
will be trying to renew contracts during the download. This also
occurs for testPeriod/2, but the test does not fail because of the
API polling immediately after the mining: the poll ends when the
contract ID changes, but when testPeriod blocks are mined, the
contract ID will have changed twice, not once.