switch to diff-based updates for tpool
Transaction pool was previously updating subscribers on the new state of the
transaction pool by sending the entire transaction pool to them every single
time a new transaction got added. This was incredibly slow, especially
considering all of the IDs and stuff that was happening behind the scenes.

Now only the diffs are sent whenever there's a transaction pool update, instead
of sending everything. This has resulted in huge speedups that should allow us
to safely increase the transaction pool from 2mb to something larger.