change interaction exponentiation
Previously the exponentiation was set to the 6th power, however there are a lot
of controls in place that prevent a bad host from being used frequently, which
means that it can take a long time for a bad host to lose enough score to be
pruned from the set of contracts.

The new exponentitaion of 15 ensures that hosts with consistently bad records
are pruned relatively quickly, even though they are not going to be used very
much.