add RPC call Metadata requesting sector IDs
This is second attempt of https://github.com/NebulousLabs/Sia/pull/2321

The whole list of IDs can be expensive for a host to provide for free,
but a short list of sector IDs (up to 2^17) is cheap. At the same time
it opens doors for stateless clients, i.e. recovering everything from
seed only.