make the chunk responsible for stanby workers
previously, workers would remember which chunks they were on standby for, and
periodically check whether they were needed to perform extra work. This had some
downtime issues, and complicated a lot of the worker functions.

Now, the chunk itself remembers which workers are on standby, and will notify
them if there's a failure that requires another worker to take action.