DownloadQueue -> DownloadHistory; Criteria carveout
This is really two commits, but I didn't separate them.

The first one is that DownloadQueue was switched to DownloadHistory, and it was
given its own mutex because it's always accessed in isolation, so there's no
reason to have blocking around the queue block the entire renter.

The second commit is that I wrote a bunch of documentation about the carveouts I
made to discriminate workers in-the-moment based on latency and price, etc, and
also shaped the code a little bit better so that those things can be easily
added later.