Display build information
When running your own build it can sometimes be difficult to identify exactly which commit you are running from.  This adds additional build information to the `siac version` command, and displays the git revision hash when `siad` starts up.  If the build was not completed via the Makefile `siad` will display a warning on start and `siac version` will not display any of the build information.

gofmt and typo