overhaul SignTransaction
In a nutshell, SignTransaction now does less work: it requires the
user to fill out the UnlockConditions and TransactionSignatures of
the transaction, whereas before it would fill them in itself. This
is a better approach because it makes the most common operation --
signing all the inputs that the wallet controls -- dead simple. And
requiring the user to fill out the unlock conditions isn't a big
deal, because they can get those from /wallet/unspent. Lastly, if
the user is responsible for filling out the TransactionSignature
fields, they can control precisely what gets signed.